# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Knowledge Melting Pot API
  version: 1.0.0
  description: Versioned API for auditable knowledge extraction and publication.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /materials:
    get:
      summary: List materials bound to a round and sub-scene for the workbench
      description: >-
        Returns safe workbench rows for every material status, including PENDING_UPLOAD, UPLOADED, SCANNING,
        FAILED and INACTIVE. Object keys, hashes, object-storage credentials, and presigned URLs are never
        included; the round must belong to the requested sub-scene.
      operationId: listMaterials
      parameters:
        - name: roundId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: subSceneId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Safe workbench material rows with their current binding.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaterialListItem'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
  /materials/upload-intents:
    post:
      summary: Declare a quarantined material upload and report upload capability
      operationId: createUploadIntent
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadIntentRequest'
      responses:
        '201':
          description: Material declaration created. No presigned URL is implied by this response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadIntent'
        '413':
          $ref: '#/components/responses/Problem'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /materials/upload-intents/{intentId}/complete:
    post:
      summary: Complete a multipart upload and queue ingest validation
      operationId: completeUpload
      parameters:
        - name: intentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteUploadRequest'
      responses:
        '202':
          $ref: '#/components/responses/AcceptedJob'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /materials/upload-intents/{intentId}:
    parameters:
      - name: intentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      summary: Abort an upload intent and its multipart upload
      operationId: abortUpload
      responses:
        '204':
          description: Upload intent aborted.
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /materials/{materialId}:
    get:
      summary: Get immutable material metadata and its round bindings
      operationId: getMaterial
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Material metadata and partitioned bindings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
        '404':
          $ref: '#/components/responses/Problem'
...
  /jobs/{jobId}:
    get:
      summary: Get durable job state
      operationId: getJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Durable job state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/Problem'
...
  /scenes/{sceneId}/release-validations:
    post:
      summary: Validate cumulative release readiness
      operationId: validateRelease
      parameters:
        - $ref: '#/components/parameters/SceneId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReleaseRequest'
      responses:
        '200':
          description: Release readiness report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseValidation'
        '412':
          $ref: '#/components/responses/Problem'
  /scenes/{sceneId}/releases/latest:
    parameters:
      - $ref: '#/components/parameters/SceneId'
    get:
      summary: Get the latest published release of a scene (the current baseline)
      description: Used to seed expectedBaseReleaseId for cumulative releases; 404 when the scene has no published release yet.
      operationId: getLatestRelease
      responses:
        '200':
          description: The latest published release metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '404':
          $ref: '#/components/responses/Problem'
  /scenes/{sceneId}/releases:
    post:
      summary: Publish an immutable cumulative scene release
      operationId: createRelease
      parameters:
        - $ref: '#/components/parameters/SceneId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReleaseRequest'
      responses:
        '201':
          description: Immutable cumulative scene release.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '409':
          $ref: '#/components/responses/Problem'
        '412':
          $ref: '#/components/responses/Problem'
  /releases/{releaseId}:
    get:
      summary: Get immutable release metadata
      operationId: getRelease
      parameters:
        - name: releaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Immutable release metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '404':
          $ref: '#/components/responses/Problem'
  /releases/{releaseId}/manifest:
    get:
      summary: Get an immutable release manifest
      operationId: getReleaseManifest
      parameters:
        - name: releaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Immutable release manifest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseManifest'
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: KMP_SESSION
  parameters:
    SceneId:
      name: sceneId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
  responses:
    AcceptedJob:
      description: Command accepted as a durable job.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AcceptedJob'
    Problem:
      description: RFC 7807 problem response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    MaterialPartition:
      type: string
      enum: [SOURCE, LABELED_TRAIN, LABELED_HOLDOUT]
    MaterialShareScope:
      type: string
      enum: [ROUND, SUBSCENE, SCENE]
    CreateUploadIntentRequest:
      type: object
      additionalProperties: false
      required: [fileName, sizeBytes, mediaType, sha256, roundId, partition, shareScope]
      properties:
        fileName: { type: string, minLength: 1, maxLength: 255 }
        sizeBytes: { type: integer, format: int64, minimum: 1, maximum: 209715200 }
        mediaType:
          type: string
          enum:
            - application/pdf
            - application/vnd.openxmlformats-officedocument.wordprocessingml.document
            - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
            - text/plain
        sha256: { type: string, pattern: '^[a-fA-F0-9]{64}$' }
        roundId: { type: string, format: uuid }
        subSceneIds:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
        partition: { $ref: '#/components/schemas/MaterialPartition' }
        shareScope: { $ref: '#/components/schemas/MaterialShareScope' }
        regulatorySource: { type: boolean, default: false }
    UploadIntent:
      type: object
      additionalProperties: false
      required:
        - id
        - materialId
        - objectKey
        - materialStatus
        - uploadMode
        - capabilityStatus
        - uploadUrlAvailable
        - maxBytes
        - supportedFormats
        - completionBehavior
        - messageCode
      properties:
        id: { type: string, format: uuid }
        materialId: { type: string, format: uuid }
        objectKey: { type: string }
        materialStatus: { type: string, enum: [PENDING_UPLOAD] }
        uploadMode:
          type: string
          enum: [DECLARATION_ONLY, MULTIPART_PRESIGNED]
        capabilityStatus:
          type: string
          enum: [OBJECT_STORAGE_NOT_CONFIGURED, MULTIPART_PRESIGNED]
        uploadUrlAvailable: { type: boolean }
        maxBytes: { type: integer, format: int64, const: 209715200 }
        supportedFormats:
          type: array
          prefixItems:
            - { type: string, const: pdf }
            - { type: string, const: docx }
            - { type: string, const: xlsx }
            - { type: string, const: txt }
          items: false
        completionBehavior: { type: string, enum: [QUEUES_VALIDATION] }
        messageCode:
          type: string
          enum:
            - material.upload.object-storage-not-configured
            - material.upload.multipart-presigned
        parts:
          type: array
          items:
            $ref: '#/components/schemas/PresignedPartMetadata'
          description: >-
            Ordered part metadata for browsers to PUT directly to object storage. Only present when
            uploadMode is MULTIPART_PRESIGNED; headers exclude the browser-managed Host header.
        partSize:
          type: [integer, 'null']
          format: int64
          minimum: 1
        partCount:
          type: [integer, 'null']
          minimum: 1
        presignedUrls:
          type: array
          items:
            type: string
            format: uri
    PresignedPartMetadata:
      type: object
      additionalProperties: false
      required: [partNumber, url]
      properties:
        partNumber: { type: integer, minimum: 1 }
        url: { type: string, format: uri }
        headers:
          type: object
          additionalProperties: { type: string }
    MaterialListItem:
      type: object
      additionalProperties: false
      required: [id, fileName, format, mediaType, sizeBytes, status, createdAt, updatedAt, binding]
      properties:
        id: { type: string, format: uuid }
        fileName: { type: string }
        format: { type: string, enum: [PDF, DOCX, XLSX, TXT] }
        mediaType: { type: string }
        sizeBytes: { type: integer, format: int64, minimum: 1, maximum: 209715200 }
        status: { $ref: '#/components/schemas/MaterialStatus' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        binding: { $ref: '#/components/schemas/MaterialBinding' }
    MaterialBinding:
      type: object
      additionalProperties: false
      required: [id, roundId, subSceneId, partition, shareScope, regulatorySource, active]
      properties:
        id: { type: string, format: uuid }
        roundId: { type: string, format: uuid }
        subSceneId: { type: string, format: uuid }
        partition: { $ref: '#/components/schemas/MaterialPartition' }
        shareScope: { $ref: '#/components/schemas/MaterialShareScope' }
        regulatorySource: { type: boolean }
        active: { type: boolean }
    CompleteUploadRequest:
      type: object
      additionalProperties: false
      properties:
        parts:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/UploadedPart'
    UploadedPart:
      type: object
      additionalProperties: false
      required: [partNumber, etag]
      properties:
        partNumber: { type: integer, minimum: 1 }
        etag: { type: string, minLength: 1, maxLength: 200 }
    MaterialStatus:
      type: string
      enum: [PENDING_UPLOAD, UPLOADED, SCANNING, READY, FAILED, INACTIVE]
    Material:
      type: object
      additionalProperties: false
      required: [id, fileName, format, mediaType, objectKey, sha256, sizeBytes, status, bindings]
      properties:
        id: { type: string, format: uuid }
        fileName: { type: string }
        format: { type: string, enum: [PDF, DOCX, XLSX, TXT] }
        mediaType: { type: string }
        objectKey: { type: string }
        sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        sizeBytes: { type: integer, format: int64, minimum: 1, maximum: 209715200 }
        status: { $ref: '#/components/schemas/MaterialStatus' }
        bindings:
          type: array
          items: { $ref: '#/components/schemas/RoundMaterial' }
    RoundMaterial:
      type: object
      additionalProperties: false
      required: [id, roundId, subSceneId, partition, shareScope, regulatorySource, active]
      properties:
        id: { type: string, format: uuid }
        roundId: { type: string, format: uuid }
        subSceneId: { type: string, format: uuid }
        partition: { $ref: '#/components/schemas/MaterialPartition' }
        shareScope: { $ref: '#/components/schemas/MaterialShareScope' }
        regulatorySource: { type: boolean }
        active: { type: boolean }
...
    AcceptedJob:
      type: object
      required: [jobId, status, statusUrl, eventsUrl]
      properties:
        jobId: { type: string, format: uuid }
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
        statusUrl: { type: string }
        eventsUrl: { type: string }
        resourceType: { type: [string, 'null'] }
        resourceId: { type: [string, 'null'], format: uuid }
    Job:
      type: object
      required: [id, type, status, attempt, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        type:
          type: string
          enum: [INGEST, EXTRACT, REEXTRACT, ALIGN, GENERATE_ASSET, GENERATE_ALL, SCENE_EXPLORE, EVALUATE]
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
        stage: { type: [string, 'null'] }
        percent: { type: integer, minimum: 0, maximum: 100 }
        attempt: { type: integer, minimum: 0 }
        errorCode: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    CreateReleaseRequest:
      type: object
      additionalProperties: false
      required: [tag, selectedSubSceneIds, note, confirmed, expectedBaseReleaseId]
      properties:
        tag: { type: string, pattern: '^v[0-9]+\.[0-9]+(?:\.[0-9]+)?$' }
        selectedSubSceneIds:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, format: uuid }
        note: { type: string, minLength: 1, maxLength: 2000 }
        expectedBaseReleaseId: { type: [string, 'null'], format: uuid }
        confirmed: { type: boolean, const: true }
    ReleaseValidation:
      type: object
      required: [ready, coverage, baseReleaseId, selected, carriedForward, missing, blockers, warnings]
      properties:
        ready: { type: boolean }
        coverage: { type: string, enum: [PARTIAL, FULL] }
        baseReleaseId: { type: [string, 'null'], format: uuid }
        selected:
          type: array
          items: { type: string, format: uuid }
        carriedForward:
          type: array
          items: { type: string, format: uuid }
        missing:
          type: array
          items: { type: string, format: uuid }
        blockers:
          type: array
          items: { type: string }
        warnings:
          type: array
          items: { type: string }
    Release:
      type: object
      required: [id, sceneId, tag, coverage, note, previousReleaseId, manifestSha256, createdAt]
      properties:
        id: { type: string, format: uuid }
        sceneId: { type: string, format: uuid }
        tag: { type: string }
        coverage: { type: string, enum: [PARTIAL, FULL] }
        note: { type: string }
        previousReleaseId: { type: [string, 'null'], format: uuid }
        manifestSha256: { type: string }
        createdAt: { type: string, format: date-time }
    ReleaseManifest:
      type: object
      description: The sha256 field hashes the canonical JSON representation with the sha256 property omitted.
      required: [schemaVersion, releaseId, sceneId, tag, coverage, note, generatedAt, previousReleaseId, subScenes, sha256]
      properties:
        schemaVersion: { type: string, const: '1.0' }
        releaseId: { type: string, format: uuid }
        sceneId: { type: string, format: uuid }
        tag: { type: string }
        coverage: { type: string, enum: [PARTIAL, FULL] }
        note: { type: string }
        generatedAt: { type: string, format: date-time }
        previousReleaseId: { type: [string, 'null'], format: uuid }
        subScenes:
          type: array
          items:
            type: object
            required: [subSceneId, status]
            properties:
              subSceneId: { type: string, format: uuid }
              status: { type: string, enum: [SELECTED, CARRIED_FORWARD, MISSING] }
              sourceReleaseId: { type: [string, 'null'], format: uuid }
              documentRevisionId: { type: [string, 'null'], format: uuid }
              assets:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  required: [assetId, assetType, assetVersion, documentRevisionId, objectKey, checksum]
                  properties:
                    assetId: { type: string, format: uuid }
                    assetType:
                      type: string
                      enum: [RULE_CATALOG, DECISION_FLOW, SKILL_PACKAGE, QA_PAIRS, EVALUATION_SET]
                    assetVersion: { type: integer, minimum: 1 }
                    documentRevisionId: { type: string, format: uuid }
                    objectKey: { type: string }
                    checksum: { type: string }
        sha256: { type: string }
...
    Problem:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }